fix: backend fixes and frontend updates
