+++ .env
@@ -1,0 +1,4 @@
+#VITE_API_BASE_URL_APP=http://localhost:8000/api
+API_BASE = /api
+VITE_API_BASE_URL = http://localhost:8000
+VITE_APP_VERSION=1.0.1
