+++ .gitignore
@@ -16,7 +16,7 @@ node_modules
 dist
 dist-ssr
 *.local
-.env
+#.env
 
 # -------------------------
 # Editor directories and files
