+++ backend/.env
@@ -1,0 +1,26 @@
+MONGO_URL = mongodb://localhost:27017/
+DB_NAME = sukanya_db
+
+
+# JWT / Security
+SECRET_KEY=12345
+ALGORITHM=HS256
+ACCESS_TOKEN_EXPIRE_MINUTES=30
+
+
+# To setup new email addres:
+
+# 1.Open Google Account: https://myaccount.google.com/
+# 2.Go to Security
+# 3.Turn on 2-Step Verification (if not already).
+# 4.After enabling 2FA, go to:
+# https://myaccount.google.com/apppasswords
+# 5.Choose:
+#         App: Mail
+#         Device: Other (Custom) → Type: FastAPI
+# 6.Google gives you a 16-character password.
+
+EMAIL_ADDRESS=sukanyayerrabati15@gmail.com
+EMAIL_PASSWORD=rsiuyhdyroywbztk
+SMTP_SERVER=smtp.gmail.com
+SMTP_PORT=587
