+++ src/context/OrderContext.tsx
@@ -1,34 +1,21 @@
 /* eslint-disable react-refresh/only-export-components */
-import { createContext, useContext, useState, useEffect, type ReactNode } from "react";
-
-export interface OrderItem {
-  name: string;
-  brand: string;
-  category: string;
-  hours: string;
-}
-
-interface OrdersContextType {
-  orders: OrderItem[];
-  addOrder: (items: CartItem[]) => void;
-}
-
-interface CartItem {
-  brand: string;
-  category: string;
-  quantity: number;
-}
+import {
+  createContext,
+  useContext,
+  useState,
+  useEffect,
+  type ReactNode,
+} from "react";
+import type { CartItem, OrderItem, OrdersContextType } from "../common/types";
 
 const OrdersContext = createContext<OrdersContextType | undefined>(undefined);
 
 export const OrdersProvider = ({ children }: { children: ReactNode }) => {
   const [orders, setOrders] = useState<OrderItem[]>(() => {
-    // ✅ Load from localStorage on first render
     const stored = localStorage.getItem("orders");
     return stored ? JSON.parse(stored) : [];
   });
 
-  // ✅ Save to localStorage whenever orders change
   useEffect(() => {
     localStorage.setItem("orders", JSON.stringify(orders));
   }, [orders]);
