+++ src/contexts/AuthContext.tsx
@@ -1,36 +1,8 @@
 /* eslint-disable react-refresh/only-export-components */
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import React, {
-  createContext,
-  useState,
-  useContext,
-  useEffect,
-  type ReactNode,
-} from "react";
+import React, { createContext, useState, useContext, useEffect } from "react";
 import axios from "axios";
-
-interface User {
-  id: string;
-  name: string;
-  email: string;
-  role: string;
-}
-
-interface AuthContextType {
-  user: User | null;
-  login: (email: string, password: string) => Promise<void>;
-  signup: (
-    name: string,
-    email: string,
-    password: string,
-    role: string
-  ) => Promise<void>;
-  logout: () => void;
-  isAuthenticated: boolean;
-  isLoading: boolean;
-  isSigningUp: boolean; 
-  isLoggingIn: boolean; 
-}
+import type { AuthContextType, AuthProviderProps, User } from "../common/types";
 
 const AuthContext = createContext<AuthContextType | undefined>(undefined);
 
@@ -41,10 +13,6 @@ export const useAuth = () => {
   }
   return context;
 };
-
-interface AuthProviderProps {
-  children: ReactNode;
-}
 
 export const AuthProvider: React.FC<AuthProviderProps> = ({ children }) => {
   const [user, setUser] = useState<User | null>(null);
@@ -116,8 +84,6 @@ export const AuthProvider: React.FC<AuthProviderProps> = ({ children }) => {
           role,
         }
       );
-
-      console.log("Signup successful:", response.data);
 
       return response.data;
     } catch (error: any) {
